delete classname,app.jsx with main.jsx in one folder,nav-link deleted,app play role Routes
