+++ src/pages/Dashboard/Dashboard.css
@@ -1,16 +1,7 @@
 #root {
   max-width: 1280px;
   margin: 0 auto;
-  /* padding: 2rem; */
   text-align: center;
-}
-
-
-/* Test Style NavLink */
-.logo_check{
-  width: 50vw;
-  height: 50vh;
-  font-size: 48px;
 }
