+++ src/pages/Dashboard/Dashboard.tsx
@@ -1,0 +1,13 @@
+// import { useState } from 'react'
+import "./Dashboard.css";
+import React from "react";
+
+function Dashboard() {
+  return (
+    <>
+      <div> Dashboard</div>
+    </>
+  );
+}
+
+export default Dashboard;
